New encoder api for input pics with custom stride

+++ test/xvc_test/decoder_helper.h
@@ -56,7 +56,7 @@ public:
   void DecodeSegmentHeaderFailed(const NalUnit &nal) {
     xvc::PicNum before_num_decoded_pics = decoder_->GetNumDecodedPics();
     size_t decoded_bytes = decoder_->DecodeNal(&nal[0], nal.size());
-    EXPECT_EQ(xvc::Decoder::kInvalidNal, decoded_bytes);
+    EXPECT_EQ(size_t(xvc::Decoder::kInvalidNal), decoded_bytes);
     EXPECT_EQ(before_num_decoded_pics, decoder_->GetNumDecodedPics());
     EXPECT_FALSE(decoder_->GetDecodedPicture(&last_decoded_picture_));
     EXPECT_EQ(nullptr, last_decoded_picture_.bytes);
@@ -76,7 +76,7 @@ public:
   void DecodePictureFailed(const NalUnit &nal) {
     xvc::PicNum before_num_decoded_pics = decoder_->GetNumDecodedPics();
     size_t decoded_bytes = decoder_->DecodeNal(&nal[0], nal.size());
-    EXPECT_EQ(xvc::Decoder::kInvalidNal, decoded_bytes);
+    EXPECT_EQ(size_t(xvc::Decoder::kInvalidNal), decoded_bytes);
     EXPECT_EQ(before_num_decoded_pics, decoder_->GetNumDecodedPics());
     EXPECT_FALSE(decoder_->GetDecodedPicture(&last_decoded_picture_));
     EXPECT_EQ(0, last_decoded_picture_.size);
